requirements.txt
flask
gunicorn
python-binance
bybit
